Updated structure; Fixed TT parser
