+++ SMParser/__init__.py
@@ -1,0 +1,1 @@
+# Empty __init__ for smparser
