+++ cli.py
@@ -1,0 +1,5 @@
+from SMParser.smparserApp import main_sm_parsing
+
+if __name__ == '__main__':
+    main_sm_parsing()
+
